requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-firestore
google-cloud-logging
google-adk 
google-genai
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-gcp-trace
opentelemetry-exporter-otlp-proto-http
langsmith
requests
